pyproject.toml
[tool.poetry]
name = "mastodon-bot-template"
version = "0.1.0"
description = ""
authors = ["Mike Birdgeneau <mike.birdgeneau@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"Mastodon.py" = "^1.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
